pyproject.toml
[project]
name = "rag-llmops"
version = "0.1.0"
description = "End-to-end implementation of RAG with LLMOps best practice"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "jsonpatch==1.33",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "pymupdf>=1.27.2.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "structlog>=25.5.0",
    "uvicorn>=0.42.0",
]
